Crate domain_test_utils

Modules

test_ethereum

Test setup shared by sp-domains-fraud-proof and domain-operator tests

test_ethereum_tx

Code in this file is copied from the frontier repository. Full path is: https://github.com/subspace/frontier/blob/1c667eb43c3d087ac66dc9ed0aa44128373f5b0a/frame/ethereum/src/mock.rs If monorepo points to new commit, this file need to be in sync.