domain_test_utils

Module test_ethereum_tx

Code in this file is copied from the frontier repository. Full path is: https://github.com/subspace/frontier/blob/1c667eb43c3d087ac66dc9ed0aa44128373f5b0a/frame/ethereum/src/mock.rs If monorepo points to new commit, this file need to be in sync.

Minor changes were made to comments, derives, visibility, and the address_build() seed size.

Structs

AccessListItem

AccountInfo

ChainId

EIP1559UnsignedTransaction

EIP2930UnsignedTransaction

LegacyTransactionMessage

LegacyUnsignedTransaction

Log

Enums

Receipt

Transaction

TransactionAction

Functions

address_build

Returns an AccountInfo struct with a deterministic address and private key. Seed zero produces an invalid private key, so we add one to the supplied seed.

contract_address

Returns the contract address for the given sender and nonce. Use U256::as_u64() to convert the nonce to u64.

Type Aliases

Block